Implement consistent map service-based proxy.
