+++ core/src/main/java/io/atomix/core/set/DistributedSetType.java
@@ -15,7 +15,7 @@
  */
 package io.atomix.core.set;
 
-import io.atomix.core.map.impl.ConsistentMapService;
+import io.atomix.core.map.impl.DefaultConsistentMapService;
 import io.atomix.core.set.impl.DelegatingDistributedSetBuilder;
 import io.atomix.core.set.impl.DistributedSetResource;
 import io.atomix.primitive.PrimitiveManagementService;
@@ -51,7 +51,7 @@ public class DistributedSetType<E> implements PrimitiveType<DistributedSetBuilde
 
   @Override
   public PrimitiveService newService(ServiceConfig config) {
-    return new ConsistentMapService();
+    return new DefaultConsistentMapService();
   }
 
   @Override
