+++ core/src/test/java/io/atomix/core/map/impl/DefaultConsistentMapServiceTest.java
@@ -15,12 +15,13 @@
  */
 package io.atomix.core.map.impl;
 
-import io.atomix.core.map.impl.ConsistentMapOperations.Get;
-import io.atomix.core.map.impl.ConsistentMapOperations.Put;
+import io.atomix.core.map.ConsistentMapType;
+import io.atomix.primitive.PrimitiveId;
+import io.atomix.primitive.service.ServiceContext;
 import io.atomix.primitive.service.impl.DefaultBackupInput;
 import io.atomix.primitive.service.impl.DefaultBackupOutput;
-import io.atomix.primitive.service.impl.DefaultCommit;
 import io.atomix.primitive.session.PrimitiveSession;
+import io.atomix.primitive.session.SessionId;
 import io.atomix.storage.buffer.Buffer;
 import io.atomix.storage.buffer.HeapBuffer;
 import io.atomix.utils.concurrent.Scheduled;
@@ -31,28 +32,32 @@ import org.junit.Test;
 
 import java.time.Duration;
 
-import static io.atomix.core.map.impl.ConsistentMapOperations.GET;
-import static io.atomix.core.map.impl.ConsistentMapOperations.PUT;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 /**
  * Consistent map service test.
  */
-public class ConsistentMapServiceTest {
+public class DefaultConsistentMapServiceTest {
 
   @Test
   @SuppressWarnings("unchecked")
   public void testSnapshot() throws Exception {
-    ConsistentMapService service = new TestConsistentMapService();
+    ServiceContext context = mock(ServiceContext.class);
+    when(context.serviceType()).thenReturn(ConsistentMapType.instance());
+    when(context.serviceName()).thenReturn("test");
+    when(context.serviceId()).thenReturn(PrimitiveId.from(1));
+    when(context.wallClock()).thenReturn(new WallClock());
 
-    service.put(new DefaultCommit<>(
-        2,
-        PUT,
-        new Put("foo", "Hello world!".getBytes(), 1000),
-        mock(PrimitiveSession.class),
-        System.currentTimeMillis()));
+    PrimitiveSession session = mock(PrimitiveSession.class);
+    when(session.sessionId()).thenReturn(SessionId.from(1));
+
+    DefaultConsistentMapService service = new TestConsistentMapService();
+    service.init(context);
+
+    service.put("foo", "Hello world!".getBytes());
 
     Buffer buffer = HeapBuffer.allocate();
     service.backup(new DefaultBackupOutput(buffer, service.serializer()));
@@ -60,19 +65,12 @@ public class ConsistentMapServiceTest {
     service = new TestConsistentMapService();
     service.restore(new DefaultBackupInput(buffer.flip(), service.serializer()));
 
-    Versioned<byte[]> value = service.get(new DefaultCommit<>(
-        2,
-        GET,
-        new Get("foo"),
-        mock(PrimitiveSession.class),
-        System.currentTimeMillis()));
+    Versioned<byte[]> value = service.get("foo");
     assertNotNull(value);
     assertArrayEquals("Hello world!".getBytes(), value.value());
-
-    assertNotNull(service.entries().get("foo").timer);
   }
 
-  private static class TestConsistentMapService extends ConsistentMapService {
+  private static class TestConsistentMapService extends DefaultConsistentMapService {
     @Override
     protected Scheduler getScheduler() {
       return new Scheduler() {
